Lots of style changes
I fixed the issue with the window having to scroll to nothing on the
right side of the browser. I also embbeded 3 google fonts in the styles
page. I made Roboto slab for body text, Black Ops One for headings, and
Audiowide for the footer. I also add boxes around different pieces of
body text to make them stand out more. I also fixed the responsiveness
of the nav bar.