Added Cascade Page to projects tab
I added a new project page for Cascade (the shipped title I worked on at
funkitron over the summer). It has info about the game, a summary of
what it is, what work i did on this project. It also has screenshots,
trailer, and links to the game download page. I also added a updated
download of my resume.