Added Beta Rangers to the Portfolio site
I have added the game Beta Rangers to the portfolio projects. Beta
Rangers now has its own dedicated game page with info about it, what I
did on it, screenshots of the game, and links to download the game and
see the git repositiory. The Project page now has Beta rangers on it,
and all of the pages nav bars have been updated to include Beta rangers.