Added Slime Spree to the Portfolio site
Slime Spree has been added to the portfolio. I has been added to the
overal projects page and all the nav bars have been updated with it. The
slime spree page includes a summary of the game, What I did on the game,
screenshots of the game, awards it has won, and a link to the github
repository.